pyproject.toml
[tool.poetry]
name = "YOHO"
version = "0.1.0"
description = "You Only Hear Once"
license = "MIT"
authors = ["JirkaKlimes <jiriklimes920@gmail.com>"]
readme = "README.md"
exclude = ["train", "experiments"]

[tool.poetry.dependencies]
python = "^3.11"
jaxlib = { version = "^0.4.26+cuda12.cudnn89", platform = "linux", source = "jax" }
flax = "^0.8.2"
ffmpeg-python = "^0.2.0"
pydantic = "^2.7.1"
toml = "^0.10.2"
sounddevice = "^0.4.6"
tqdm = "^4.66.4"
tabulate = "^0.9.0"
sentencepiece = "^0.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pre-commit = "^3.7.0"
ruff = "^0.4.4"
num2words = "^0.5.13"
tiktoken = "^0.6.0"
matplotlib = "^3.8.4"
pyqt6 = "^6.7.0"
googletrans = "^3.1.0a0"
eld = "^1.0.6"
audiomentations = "^0.35.0"
pydantic = "^2.7.1"
srt = "^3.5.3"

[[tool.poetry.source]]
name = "jax"
url = "https://storage.googleapis.com/jax-releases/jax_cuda_releases.html"
priority = "primary"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
indent-width = 4
target-version = "py311"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
